+++ tabs/base.py
@@ -10,6 +10,7 @@
 import logging
 
 import wx
+import browsableObjects
 import clipboardHelper
 import errorCodes
 import globalVars
@@ -260,4 +261,35 @@ class FalconTabBase(object):
 
 	def Move(self,path,cursor=""):
 		"""指定の場所へ移動する。"""
-		return navigator.Navigate(path,cursor,self)
+		r=navigator.Navigate(path,cursor,previous_tab=self)
+		return errorCodes.OK if r is self else r
+
+	def GoForward(self,stream,admin=False):
+		"""選択中のフォルダに入るか、選択中のファイルを実行する。stream=True の場合、ファイルの NTFS 副ストリームを開く。"""
+		index=self.GetFocusedItem()
+		elem=self.listObject.GetElement(index)
+		if isinstance(elem,browsableObjects.Folder):#このフォルダを開く
+			#TODO: 管理者モードだったら、別のfalconが昇格して開くように
+			return self.Move(elem.fullpath)
+		#end フォルダ開く
+		elif isinstance(elem,browsableObjects.File):#このファイルを開く
+			if not stream:
+				misc.RunFile(elem.fullpath,admin)
+				return
+			#end runFile
+			#TODO: 管理者として副ストリーム…まぁ、使わないだろうけど一貫性のためには開くべきだと思う
+			if stream:
+				return navigator.Navigate(elem.fullpath)
+		#end ファイルを開く
+		#end なにを開くか
+	#end GoForward
+
+	def GoBackward(self):
+		"""内包しているフォルダ/ドライブ一覧へ移動する。"""
+		if len(self.listObject.rootDirectory)<=3:		#ドライブリストへ
+			target=""
+			cursorTarget=self.listObject.rootDirectory[0]
+		else:
+			target=os.path.split(self.listObject.rootDirectory)[0]
+			cursorTarget=os.path.split(self.listObject.rootDirectory)[1]
+		return self.move(target,cursorTarget)
